restructure the go
