+++ services/deal/storageDeal.go
@@ -1,0 +1,50 @@
+package deal
+
+import (
+	"context"
+	"log"
+	"time"
+
+	"github.com/FIL-Builders/xchainClient/config"
+)
+
+// SmartContractDeal continuously runs logic until the context is canceled
+func SmartContractDeal(ctx context.Context, cfg *config.Config, srcCfg *config.SourceChainConfig) error {
+	log.Println("Starting SmartContractDeal process...")
+
+	// Example: Perform a periodic task in a loop until the context is canceled
+	ticker := time.NewTicker(5 * time.Second) // Adjust the interval as needed
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			// Handle graceful shutdown
+			log.Println("SmartContractDeal process is shutting down...")
+			return nil
+		case <-ticker.C:
+			// Example logic: Interact with a smart contract or process data
+			err := processSmartContractLogic(cfg, srcCfg)
+			if err != nil {
+				log.Printf("Error in SmartContractDeal process: %v", err)
+			} else {
+				log.Println("SmartContractDeal task completed successfully")
+			}
+		}
+	}
+}
+
+// processSmartContractLogic handles the core logic of the SmartContractDeal process
+func processSmartContractLogic(cfg *config.Config, srcCfg *config.SourceChainConfig) error {
+	// Example placeholder logic:
+	// This is where you could interact with a smart contract, fetch data, or perform computations.
+
+	log.Printf("Processing smart contract logic with chain: %s and config: %+v", srcCfg.OnRampAddress, cfg)
+
+	// TODO: Implement actual logic to interact with a smart contract
+	// - Call an Ethereum or Filecoin smart contract
+	// - Process events or data
+	// - Execute business logic
+
+	return nil // Return an error if something fails
+}
